Correction actualisation

+++ front/groupomania/src/components/Profil/Profile.js
@@ -28,32 +28,33 @@ const ProfileForm = () => {
 	}, [uid]);
 
 	const deleteProfile = () => {
-		window.confirm(
-			'Êtes-vous sûr de vouloir supprimer votre compte ? Cette action est irréversible.'
-		);
-		axios({
-			method: 'delete',
-			url: `${process.env.REACT_APP_API_URL}api/user/${uid}`,
-			withCredentials: true,
-		})
-			.then((res) => {
-				if (res.data.errors) {
-					console.log(res.data.errors);
-				} else {
-					console.log(res);
-					window.location = '/';
-				}
+		if (
+			window.confirm(
+				'Êtes-vous sûr de vouloir supprimer votre compte ? Cette action est irréversible.'
+			) === true
+		) {
+			axios({
+				method: 'delete',
+				url: `${process.env.REACT_APP_API_URL}api/user/${uid}`,
+				withCredentials: true,
 			})
-			.catch((err) => {
-				console.log(err);
-			});
+				.then((res) => {
+					if (res.data.errors) {
+						console.log(res.data.errors);
+					} else {
+						console.log(res);
+						window.location = '/';
+					}
+				})
+				.catch((err) => {
+					console.log(err);
+				});
+		} else {
+			return;
+		}
 	};
 
 	const editProfile = () => {
-		// const formPicture = new FormData();
-		// formPicture.append('id', uid);
-		// formPicture.append('email', email);
-		// formPicture.append('pseudo', pseudo);
 		axios({
 			method: 'put',
 			url: `${process.env.REACT_APP_API_URL}api/user/${uid}`,
